+++ src/hooks/account/useGetAccountFromApi.ts
@@ -1,33 +1,25 @@
-import axios from 'axios';
+import { useEffect } from 'react';
 import useSwr from 'swr';
 
-import { ACCOUNTS_ENDPOINT } from 'apiCalls';
-import { apiAddressSelector } from 'reduxStore/selectors';
-import { store } from 'reduxStore/store';
+import { accountFetcher } from 'apiCalls/accounts/getAccountFromApi';
 
-export interface UseGetAccountFromApiOptionsType {
-  shouldSkipFetching?: boolean;
-}
-
-export const useGetAccountFromApi = (
-  address: string,
-  options: UseGetAccountFromApiOptionsType
-) => {
-  const apiAddress = apiAddressSelector(store.getState());
-  const accountEndpoint = options.shouldSkipFetching
-    ? null
-    : `${apiAddress}/${ACCOUNTS_ENDPOINT}/${address}`;
-
-  const { data, error, isLoading, isValidating } = useSwr(
-    accountEndpoint,
-    (url) => axios.get(url).then((response) => response.data)
+export const useGetAccountFromApi = (address: string | null) => {
+  const { data, error, isLoading, isValidating, mutate } = useSwr(
+    address,
+    (address) => accountFetcher(address).then((response) => response.data),
+    { revalidateOnFocus: false, revalidateOnMount: false }
   );
 
-  const dataLoadingFromLibrary = isLoading || isValidating;
-  const isUsernameLoading = dataLoadingFromLibrary || (!error && !data);
+  const startFetchOnMount = () => {
+    if (!data) {
+      mutate();
+    }
+  };
+
+  useEffect(startFetchOnMount, [data, mutate]);
 
   return {
-    loading: isUsernameLoading && !options.shouldSkipFetching,
+    loading: isLoading || isValidating,
     error,
     account: data
   };